use local gulp if eligible
If gulp is in the `package.json` dependencies (`node_modules` in UI
directory) then it is preferred over using the `$PATH` to resolve gulp.
This makes the gulp tasks "just work"™ after NPM install without
requiring the user to either install gulp globally or add the
`node_modules` to the path.