add semaphore to ensure gulp watch doesn't run twice concurrently
For some reason the hook is run twice when `sbt run` is executed,
which results in two concurrent `gulp watch` processes that create a
lot of issues (concurrent prettify leads to truncated JS files,
deleted JS files, etc.).
This should be fixed in playframework upstream (I assume the issue
stems from there because the hook implementation is a textbook example
from the play documentation) but it is not really difficult to ensure
that the watcher is run only once anyway.