Drops content-disposition header
The Content-Disposition header breaks the html import feature [1] which is
commonly used by web components frameworks such as polymer.
Therefore the `Ok.sendfile` helper is not appropriate for `GulpAssets` and
should be reserved to send over files to be downloaded
(whether forced downlading when disposition is attachment or providing a
filename for a preview when disposition is inline)

This commits drops `Ok.sendfile` in favor of `RangeResult.ofSource` which is the
method used by playframework's onw `Assets` controller.

I upgraded to play 2.5.4 to have access to RangeResult which is used in the
current (2.5.10) `Assets` controller.

I considered upgrading as far as 2.5.10 where FileIO.fromFile is deprecated but
this would force users of the plugin to upgrade too. I thought 2.5.4 was a small
enough step to be worth it.

[1] see https://www.w3.org/TR/html-imports/ and the following issues for the
rationale
 * https://bugs.chromium.org/p/chromium/issues/detail?id=439877
 * https://www.w3.org/Bugs/Public/show_bug.cgi?id=27550