Drops content-disposition header
The Content-Disposition header breaks the html import feature [1] which is
commonly used by web components frameworks such as polymer.
Therefore the Ok.sendfile helper is not appropriate for GulpAssets and
should be reserved to send over files to be downloaded
(whether forced downlading when disposition is attachment or providing a
filename for a preview when disposition is inline)

This commits drops Ok.sendfile in favor of Ok.sendEntity which is the
method used by playframework's onw Assets controller.

This change is compatible with play 2.5.0 and upward.

[1] see https://www.w3.org/TR/html-imports/ and the following issues for the
rationale
 * https://bugs.chromium.org/p/chromium/issues/detail?id=439877
 * https://www.w3.org/Bugs/Public/show_bug.cgi?id=27550